new JWT authentication implementation

+++ src/main/java/hu/student/projlab/mealride_api/config/security/SecurityConstants.java
@@ -3,7 +3,7 @@ package hu.student.projlab.mealride_api.config.security;
 public class SecurityConstants {
 
     public static final String SECRET = "SecretMealRideKeyGenerator";
-    public static final long EXPIRATION_TIME = 3_600_000; // 1 hour
+    public static final int EXPIRATION_TIME = 3_600_000; // 1 hour
     public static final String TOKEN_PREFIX = "Bearer ";
     public static final String HEADER_STRING = "Authorization";
     public static final String SIGN_UP_URL = "/sign-up";
